+++ lib/api/folder-test.ts
@@ -8,7 +8,7 @@ import {MonitorBindingController} from '../plugin/blade/common/controller/monito
 import {Blade} from '../plugin/blade/common/model/blade';
 import {FolderController} from '../plugin/blade/folder/controller';
 import {SeparatorController} from '../plugin/blade/separator/controller';
-import {Color} from '../plugin/common/model/color';
+import {Color} from '../plugin/input-bindings/color/model/color';
 import {NumberTextController} from '../plugin/input-bindings/number/controller/number-text';
 import {SingleLogMonitorController} from '../plugin/monitor-bindings/common/controller/single-log';
 import {FolderApi} from './folder';
