+++ lib/pane/input-test.ts
@@ -3,8 +3,8 @@ import {describe as context, describe, it} from 'mocha';
 
 import Tweakpane from '../index';
 import {TestUtil} from '../misc/test-util';
-import {Color} from '../plugin/common/model/color';
 import {PaneError} from '../plugin/common/pane-error';
+import {Color} from '../plugin/input-bindings/color/model/color';
 
 function createPane(): Tweakpane {
 	return new Tweakpane({
