+++ lib/plugin/input-bindings/color/model/color-model-test.ts
