+++ lib/plugin/input-bindings/color/model/color-test.ts
