+++ lib/plugin/input-bindings/color/view/color-picker.ts
@@ -1,11 +1,11 @@
 import {forceCast} from '../../../../misc/type-util';
-import {Color} from '../../../common/model/color';
 import {Foldable} from '../../../common/model/foldable';
-import {PickedColor} from '../../../common/model/picked-color';
 import {Value} from '../../../common/model/value';
 import {ClassName} from '../../../common/view/class-name';
 import {ValueView} from '../../../common/view/value';
 import {TextView} from '../../common/view/text';
+import {Color} from '../model/color';
+import {PickedColor} from '../model/picked-color';
 import {APaletteView} from './a-palette';
 import {ColorComponentTextsView} from './color-component-texts';
 import {HPaletteView} from './h-palette';
