+++ src/doc/ts/panepaint.ts
@@ -1,8 +1,8 @@
-import {Color} from '../../../lib/plugin/common/model/color';
 import {
 	colorFromString,
 	colorToFunctionalRgbaString,
 } from '../../../lib/plugin/input-bindings/color/converter/color-string';
+import {Color} from '../../../lib/plugin/input-bindings/color/model/color';
 import {createTheme, ThemeId} from './themes';
 
 declare let Tweakpane: any;
