+++ src/doc/ts/route/index.ts
@@ -1,8 +1,8 @@
-import {Color} from '../../../../lib/plugin/common/model/color';
 import {
 	colorFromString,
 	colorToFunctionalRgbaString,
 } from '../../../../lib/plugin/input-bindings/color/converter/color-string';
+import {Color} from '../../../../lib/plugin/input-bindings/color/model/color';
 import {Sketch} from '../sketch';
 import {Environment} from '../sketch';
 import {selectContainer} from '../util';
